use ToggleButton for /explore actions button; add icon for browse button
