+++ pages/etherpad/[[...roomId]].js
@@ -14,7 +14,7 @@ import { ServiceTable } from '../../components/UI/ServiceTable';
 import Form from '../../components/UI/Form';
 import LoadingSpinnerInline from '../../components/UI/LoadingSpinnerInline';
 import LoadingSpinner from '../../components/UI/LoadingSpinner';
-import ServiceIframeHeader from './ServiceIframeHeader';
+import ServiceIframeHeader from '../../components/UI/ServiceIframeHeader';
 
 export default function Etherpad() {
     const auth = useAuth();
